change ikala window to gradle structure
