+++ Ikala-Window/gradle.properties
@@ -1,0 +1,8 @@
+# Dependency versions
+junitVersion=5.4.0
+log4jAPIVersion=2.17.2
+log4jCoreVersion=2.17.2
+log4jSlf4jVersion=2.17.2
+lombokVersion=1.18.24
+slf4jVersion=1.7.36
+snakeYamlVersion=1.30
