--- Ikala-Window/src/com/ikalagaming/window/WindowPlugin.java
@@ -1,17 +1,0 @@
-package com.ikalagaming.window;
-
-import com.ikalagaming.plugins.Plugin;
-
-/**
- * An example plugin to demonstrate how plugins are set up.
- *
- * @author Ches Burks
- *
- */
-public class WindowPlugin extends Plugin {
-	/**
-	 * The name of the plugin in Java for convenience, should match the name in
-	 * plugin.yml.
-	 */
-	public static final String PLUGIN_NAME = "Ikala-Window";
-}
