+++ Ikala-Window/src/main/resources/plugin.yml
@@ -1,0 +1,5 @@
+name: Ikala-Window
+version: 0.0.1
+description: A generic window system
+authors: [maceris]
+main-class: com.ikalagaming.window.WindowPlugin
